+++ tests/TestCase/Mailer/MailerTest.php
@@ -76,6 +76,6 @@ class MailerTest extends TestCase
         $this->assertEquals(['email@example.com' => 'MeCms'], $email->from());
         $this->assertEquals('html', $email->emailFormat());
 
-        $this->assertEquals(['ipAddress' => false], $email->viewVars);
+        $this->assertEquals([], $email->viewVars);
     }
 }