+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # 2.x branch
+## 2.16 branch
+### 2.16.0
+* it uses `CollectionInterface::first()` instead of the global `firstValue()`;
+* added `StaticPage::slug()` method. Renamed `StaticPage::_getPaths()` as
+    `StaticPage::paths()`;
+* renamed `/ContactForm.php` as `ContactUsForm` and `ContactFormMailer` as
+    `ContactUsMailer`. This creates less confusion with the classes names;
+* simplified the name of some methods of the `BaseUpdateShell` class.
+
 ## 2.15 branch
+### 2.15.4
+* improved code for `BaseUpdateShell` class.
+
 ### 2.15.3
 * fixed code for `BackupForm` class. Added `_getBackupExportInstance()` method;
 * fixed code and a little bug for `ContactForm` and `ContactFormMailer` classes;